v3.30.16.3 - correcting reading/writing tag groups

+++ src/mainwindow.cpp
@@ -299,7 +299,7 @@ void MainWindow::readSettings()
     int xdelta = 0;
     QScreen *screen = QGuiApplication::primaryScreen();
     QRect scrRect = screen->availableGeometry();
-    QRect deskRect = desk->screenGeometry();
+//    QRect deskRect = desk->screenGeometry();
     QSettings settings(app_copyright_org, app_name);
     QString model (app_dir + "/default");
     QString exe (app_dir + "/ss.exe");
@@ -1007,44 +1007,45 @@ void MainWindow::runConversion()
         if (!QFileInfo(ss_trans_exe).exists())
             locateSSConverter();
     }
-    drun.setExe(ss_trans_exe);
-    drun.setOptions("-nohess");
-    drun.exec();
-    // reset last estimation phase
-    modelData->set_last_estim_phase(lastEstPhase);
-    files->write_starter_file();
-    // ask user for new directory and copy files there
+    if (QFileInfo(ss_trans_exe).exists())
     {
-    // save old directory
-    QString old_dir (current_dir);
-    // open new directory and move there
-    openNewDirectory();
-    if (old_dir == current_dir)
-    {
-        int btn =
-        QMessageBox::question(this, tr("Same Directory"), tr("You may not select the same directory.\nDo you wish to try again?"), QMessageBox::Yes, QMessageBox::No);
-        if (btn == QMessageBox::Yes)
-            openNewDirectory();
-    }
-    if (old_dir != current_dir)
-    {
-        // copy files
-        copy_file (old_dir + QString("/starter.ss_new"), current_dir + QString("/starter.ss"));
-        files->read_starter_file("starter.ss");
-        copy_file (old_dir + QString("/forecast.ss_new"), current_dir + QString("/forecast.ss"));
-        copy_file (old_dir + QString("/data.ss_new"), current_dir + QString("/") + files->getDataFileName());
-        copy_file (old_dir + QString("/control.ss_new"), current_dir + QString("/") + files->getControlFileName());
-        copy_file (old_dir + QString("/ss.par"), current_dir + QString("/ss.par"));
-        copy_file (old_dir + QString("/ss.bar"), current_dir + QString("/ss.bar"));
-        if (QFile(old_dir + QString("/wtatage.ss")).exists())
-        {
-            copy_file (old_dir + QString("/wtatage.ss_new"), current_dir + QString("/wtatage.ss"));
-        }
-        // read data into GUI
-        readFiles();
+        drun.setExe(ss_trans_exe);
+        drun.setOptions("-nohess");
+        drun.exec();
+        // reset last estimation phase
         modelData->set_last_estim_phase(lastEstPhase);
         files->write_starter_file();
-    }
+        // ask user for new directory and copy files there
+        // save old directory
+        QString old_dir (current_dir);
+        // open new directory and move there
+        openNewDirectory();
+        if (old_dir == current_dir)
+        {
+            int btn =
+            QMessageBox::question(this, tr("Same Directory"), tr("You may not select the same directory.\nDo you wish to try again?"), QMessageBox::Yes, QMessageBox::No);
+            if (btn == QMessageBox::Yes)
+                openNewDirectory();
+        }
+        if (old_dir != current_dir)
+        {
+            // copy files
+            copy_file (old_dir + QString("/starter.ss_new"), current_dir + QString("/starter.ss"));
+            files->read_starter_file("starter.ss");
+            copy_file (old_dir + QString("/forecast.ss_new"), current_dir + QString("/forecast.ss"));
+            copy_file (old_dir + QString("/data.ss_new"), current_dir + QString("/") + files->getDataFileName());
+            copy_file (old_dir + QString("/control.ss_new"), current_dir + QString("/") + files->getControlFileName());
+            copy_file (old_dir + QString("/ss.par"), current_dir + QString("/ss.par"));
+            copy_file (old_dir + QString("/ss.bar"), current_dir + QString("/ss.bar"));
+            if (QFile(old_dir + QString("/wtatage.ss")).exists())
+            {
+                copy_file (old_dir + QString("/wtatage.ss_new"), current_dir + QString("/wtatage.ss"));
+            }
+            // read data into GUI
+            readFiles();
+            modelData->set_last_estim_phase(lastEstPhase);
+            files->write_starter_file();
+        }
     }
 }
